+++ packages/mediation/gma_mediation_ironsource/android/src/main/kotlin/io/flutter/plugins/googlemobileads/mediation/gma_mediation_ironsource/GmaMediationIronsourcePlugin.kt
@@ -14,14 +14,31 @@
 
 package io.flutter.plugins.googlemobileads.mediation.gma_mediation_ironsource
 
+import com.ironsource.mediationsdk.IronSource;
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 
-/** GmaMediationIronsourcePlugin. */
-class GmaMediationIronsourcePlugin: FlutterPlugin {
-  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+/** GmaMediationIronsourcePlugin manages IronSourcePrivacyApi and implements the needed methods. */
+class GmaMediationIronsourcePlugin: FlutterPlugin, IronSourcePrivacyApi {
 
+  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    IronSourcePrivacyApi.setUp(flutterPluginBinding.binaryMessenger, this)
   }
 
   override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+    IronSourcePrivacyApi.setUp(binding.binaryMessenger, null)
+  }
+
+  override fun setConsent(gdprConsent: Boolean) {
+    IronSource.setConsent(gdprConsent)
+  }
+
+  override fun setDoNotSell(doNotSell: Boolean) {
+    IronSource.setMetaData(IRONSOURCE_DONOTSELL_KEY, if (doNotSell) TRUE else FALSE)
+  }
+
+  companion object {
+    const val IRONSOURCE_DONOTSELL_KEY = "do_not_sell"
+    const val TRUE = "true"
+    const val FALSE = "false"
   }
 }