+++ packages/mediation/gma_mediation_ironsource/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 1.0.0
+## ironSource Flutter Mediation Adapter Changelog
 
+#### 1.0.0
 * Initial release.
-* Verified compatibility with ironSource Android adapter version 7.9.0.0
-* Verified compatibility with ironSource iOS adapter version 7.9.1.0.0
+* Verified compatibility with ironSource Android adapter version 8.0.0.0
+* Verified compatibility with ironSource iOS adapter version 8.0.0.0.0
+* Built and tested with the Google Mobile Ads Flutter Plugin version 5.1.0.